fix up node issues
