+++ beacon-chain/powchain/deposit_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/signing"
 	testDB "github.com/prysmaticlabs/prysm/beacon-chain/db/testing"
+	testing2 "github.com/prysmaticlabs/prysm/beacon-chain/powchain/testing"
 	fieldparams "github.com/prysmaticlabs/prysm/config/fieldparams"
 	"github.com/prysmaticlabs/prysm/config/params"
 	"github.com/prysmaticlabs/prysm/container/trie"
@@ -52,7 +53,7 @@ func TestDepositContractAddress_OK(t *testing.T) {
 
 func TestProcessDeposit_OK(t *testing.T) {
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
@@ -80,7 +81,7 @@ func TestProcessDeposit_OK(t *testing.T) {
 
 func TestProcessDeposit_InvalidMerkleBranch(t *testing.T) {
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
@@ -110,7 +111,7 @@ func TestProcessDeposit_InvalidMerkleBranch(t *testing.T) {
 func TestProcessDeposit_InvalidPublicKey(t *testing.T) {
 	hook := logTest.NewGlobal()
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
@@ -150,7 +151,7 @@ func TestProcessDeposit_InvalidPublicKey(t *testing.T) {
 func TestProcessDeposit_InvalidSignature(t *testing.T) {
 	hook := logTest.NewGlobal()
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
@@ -189,7 +190,7 @@ func TestProcessDeposit_InvalidSignature(t *testing.T) {
 func TestProcessDeposit_UnableToVerify(t *testing.T) {
 	hook := logTest.NewGlobal()
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
@@ -225,7 +226,7 @@ func TestProcessDeposit_UnableToVerify(t *testing.T) {
 
 func TestProcessDeposit_IncompleteDeposit(t *testing.T) {
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
@@ -291,7 +292,7 @@ func TestProcessDeposit_IncompleteDeposit(t *testing.T) {
 
 func TestProcessDeposit_AllDepositedSuccessfully(t *testing.T) {
 	beaconDB := testDB.SetupDB(t)
-	server, endpoint := setupRPCServer(t)
+	server, endpoint, _ := testing2.SetupRPCServer()
 	t.Cleanup(func() {
 		server.Stop()
 	})
